Pedir confirmacion antes de eliminar evento

+++ src/pages/timeline.js
@@ -1,3 +1,4 @@
+import swal from 'sweetalert';
 import { getEvents, editEvent, deletePost } from '../firebase/post';
 
 const event = (evento) => {
@@ -45,29 +46,47 @@ const event = (evento) => {
     </div>
   `;
 
-  eventContainer.querySelector('.flaticon-menu').addEventListener('click', () => eventContainer.querySelector('ul').classList.toggle('hide'));
+  eventContainer
+    .querySelector('.flaticon-menu')
+    .addEventListener('click', () => eventContainer.querySelector('ul').classList.toggle('hide'));
   // funcion asistire
-  eventContainer.querySelector('.flaticon-strong').addEventListener('click', () => {
-    let likes = evento.likes || [];
-    if (likes.includes(user)) {
-      likes = likes.filter(like => like !== user);
-    } else {
-      likes.push(user);
-    }
-    editEvent(evento.eventId, { likes });
-    evento.likes = likes;
-    eventContainer.querySelector('.interaction__text').innerHTML = `${likes.length} Asistiré`;
-  });
+  eventContainer
+    .querySelector('.flaticon-strong')
+    .addEventListener('click', () => {
+      let likes = evento.likes || [];
+      if (likes.includes(user)) {
+        likes = likes.filter(like => like !== user);
+      } else {
+        likes.push(user);
+      }
+      editEvent(evento.eventId, { likes });
+      evento.likes = likes;
+      eventContainer.querySelector('.interaction__text').innerHTML = `${likes.length} Asistiré`;
+    });
   // funcion eliminar evento
-  eventContainer.querySelector('.delete').addEventListener('click', async () => {
-    if (user === evento.id) {
-      await deletePost(evento.eventId);
-      // console.log(evento.eventId);
-      eventContainer.innerHTML = '';
-    } else {
-      console.log('No puedes eliminar este evento');
-    }
-  });
+  eventContainer
+    .querySelector('.delete')
+    .addEventListener('click', async () => {
+      if (user === evento.id) {
+        swal({
+          title: '¿Estas seguro?',
+          text: 'Una vez eliminado, no podras recuperar este Evento',
+          icon: 'warning',
+          buttons: true,
+          dangerMode: true,
+        }).then((willDelete) => {
+          if (willDelete) {
+            deletePost(evento.eventId);
+            swal('Tu Evento ha sido eliminado', {
+              icon: 'success',
+            });
+            eventContainer.innerHTML = '';
+          }
+        });
+      } else {
+        swal('No puedes eliminar este evento');
+      }
+    });
   // funcion editar evento
   eventContainer.querySelector('.edit').addEventListener('click', async () => {
     if (user === evento.id) {
